handle a case where response can be None instead of empty

+++ ami-delete.sh
@@ -16,14 +16,14 @@ for r in ${regions[@]}; do
     snap_amd=$(AWS_REGION=$r aws ec2 describe-images --owners self --filters Name=name,Values=oyster_amd64 --no-paginate --query 'Images[0].BlockDeviceMappings[0].Ebs.SnapshotId' --output text)
     ami_arm=$(AWS_REGION=$r aws ec2 describe-images --owners self --filters Name=name,Values=oyster_arm64 --no-paginate --query 'Images[0].ImageId' --output text)
     snap_arm=$(AWS_REGION=$r aws ec2 describe-images --owners self --filters Name=name,Values=oyster_arm64 --no-paginate --query 'Images[0].BlockDeviceMappings[0].Ebs.SnapshotId' --output text)
-    if [[ $ami_amd != null ]]; then
+    if [[ $ami_amd != null && "$ami_amd" != "None" ]]; then
         echo "Deleting amd64 image in region $r: $ami_amd"
         AWS_REGION=$r aws ec2 deregister-image --image-id $ami_amd
         AWS_REGION=$r aws ec2 delete-snapshot --snapshot-id $snap_amd
     else
         echo "Found no amd64 image in $r"
     fi
-    if [[ $ami_arm != null ]]; then
+    if [[ $ami_arm != null && "$ami_arm" != "None" ]]; then
         echo "Deleting arm64 image in region $r: $ami_arm"
         AWS_REGION=$r aws ec2 deregister-image --image-id $ami_arm
         AWS_REGION=$r aws ec2 delete-snapshot --snapshot-id $snap_arm
